+++ website/src/components/Header/Header.js
@@ -16,10 +16,23 @@ export default {
     return {
       menu: Menu,
       logo: this.magic ? (this.theme === 'dark' ? LogoDark : LogoLight) : (this.theme === 'light' ? LogoDark : LogoLight),
-      quickOrderVisible: false
+      quickOrderVisible: false,
+      hasBackground: false,
+      path: '/'
+    }
+  },
+  watch: {
+    '$route.path': (next) => {
+      this.path = next
     }
   },
   mounted () {
+    const { path } = this.$route
+    this.path = path
+    document.addEventListener('scroll', this.onScroll)
+  },
+  beforeDestroy () {
+    document.removeEventListener('scroll')
   },
   methods: {
     handleMenuItemClick (item) {
@@ -29,6 +42,12 @@ export default {
           path: to
         })
       }
+    },
+    onScroll () {
+      // 滚动时候的情况
+      const top = window.scrollY
+      this.hasBackground = top > 100
+      console.log(top)
     }
   }
 }