+++ website/src/views/Guide/index.vue
@@ -1,0 +1,5 @@
+<template src="./Guide.html"></template>
+<script src="./Guide.js"></script>
+<style src="./Guide.less" lang="less" scoped>
+
+</style>