+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 2022-09-29
+- Vue3(vue-next) project version 1.0.1 published
+- update guide page for website
+
 ## 2022-09-28
 - Vue2(vue) project version 1.0.3 published
 - update icon page for website